+++ examples/deploy-contract/index.js
@@ -2,11 +2,17 @@ require('dotenv').config();
 
 const { Defender } = require('@openzeppelin/defender-sdk');
 const { AbiCoder } = require('ethers');
+const https = require('https');
 
 const artifactFile = require('./artifacts/Box.json');
 
 async function main() {
-  const creds = { apiKey: process.env.API_KEY, apiSecret: process.env.API_SECRET };
+  const creds = {
+    apiKey: process.env.API_KEY,
+    apiSecret: process.env.API_SECRET,
+    //optional https config to keep connection alive. You can pass any configs that are accepted by https.Agent
+    httpsAgent: https.Agent({ keepAlive: true }),
+  };
   const client = new Defender(creds);
 
   // await client.deploy.createBlockExplorerApiKey({