+++ packages/deploy/src/models/index.ts
@@ -3,5 +3,4 @@ export type Address = string;
 export * from './approval-process';
 export * from './upgrade';
 export * from './deployment';
-export * from './deployment-config';
 export * from './block-explorer';