+++ packages/relay-signer/src/action/index-rate.test.ts
@@ -1,8 +1,13 @@
 import { ActionRelayer } from '.';
 import Lambda from '../__mocks__/aws-sdk/clients/lambda';
+import { Lambda as LambdaV3 } from '../__mocks__/@aws-sdk/client-lambda';
+jest.mock('node:process', () => ({
+  ...jest.requireActual('node:process'),
+  version: 'v16.0.3',
+}));
 
 type TestActionRelayer = Omit<ActionRelayer, 'lambda' | 'relayerARN'> & {
-  lambda: ReturnType<typeof Lambda>;
+  lambda: ReturnType<typeof Lambda | typeof LambdaV3>;
   arn: string;
 };
 
@@ -32,13 +37,15 @@ describe('ActionRelayer', () => {
   let relayer: TestActionRelayer;
 
   beforeEach(async function () {
-    relayer = (new ActionRelayer({
+    relayer = new ActionRelayer({
       credentials: JSON.stringify(credentials),
       relayerARN: 'arn',
-    }) as unknown) as TestActionRelayer;
+    }) as unknown as TestActionRelayer;
   });
 
   describe('get rate limited', () => {
+    jest.mock('aws-sdk/clients/lambda', () => Lambda);
+    jest.mock('@aws-sdk/client-lambda', () => ({ Lambda: LambdaV3 }));
     test('throw Rate limit error after 300 requests', async () => {
       const rateLimit = 300;