+++ packages/relay-signer/src/models/transactions.ts
@@ -11,6 +11,7 @@ export interface SendBaseTransactionRequest {
   gasLimit: BigUInt;
   validUntil?: string;
   isPrivate?: boolean;
+  privateMode?: PrivateTransactionMode;
 }
 
 export interface SendSpeedTransactionRequest extends SendBaseTransactionRequest {
@@ -32,6 +33,14 @@ export type RelayerTransactionPayload =
   | SendLegacyTransactionRequest
   | SendEIP1559TransactionRequest;
 
+export type PrivateTransactionMode = FlashbotTransactionMode;
+/**
+ * Fast mode has 2 key differences from the default Protect experience:
+ * 1. Shared with all builders: By default, Protect transactions are only shared with the Flashbots Builder, which builds only a subset of all Ethereum blocks. In fast mode, transactions are shared with all registered builders to increase the number of blocks the user's transaction can be included in.
+ * 2. Larger refund paid to validator: By default, only 10% of MEV-Share refunds are paid to validators. In fast mode, validators receive 50% of refunds which makes it more likely that the user’s transactions will be chosen in a given block.
+ */
+export type FlashbotTransactionMode = 'flashbots-normal' | 'flashbots-fast';
+
 interface RelayerTransactionBase {
   transactionId: string;
   hash: string;
@@ -49,6 +58,12 @@ interface RelayerTransactionBase {
   sentAt?: string;
   pricedAt?: string;
   isPrivate?: boolean;
+  privateMode?: PrivateTransactionMode;
+  signature?: {
+    v: string;
+    r: string;
+    s: string;
+  };
 }
 
 interface RelayerLegacyTransaction extends RelayerTransactionBase {