+++ examples/update-action-env-variables/CHANGELOG.md
@@ -1,5 +1,81 @@
 # @openzeppelin/defender-sdk-example-update-action
 
+## 1.13.1
+
+### Patch Changes
+
+- e55d50a: Patch: Fix aws-sdk v3 action error payload
+- Updated dependencies [e55d50a]
+  - @openzeppelin/defender-sdk@1.13.1
+
+## 1.13.0
+
+### Minor Changes
+
+- ff4a967: - feat: add exponential backoff logic for retry
+  - fix: add conditional imports for v2 and v3 for actions
+  - chore: add loadbalance example for rate limit number of transactions
+
+### Patch Changes
+
+- Updated dependencies [ff4a967]
+  - @openzeppelin/defender-sdk@1.13.0
+
+## 1.12.0
+
+### Minor Changes
+
+- 6b45bce: - feat: Support Holesky & Amoy networks.
+  - chore: Add retries to 520 errors.
+  - feat: Allow optional httpsAgent options.
+  - fix: Avoid relaySigner ENS resolution by default.
+  - chore: Upgrade dependencies.
+
+### Patch Changes
+
+- Updated dependencies [6b45bce]
+  - @openzeppelin/defender-sdk@1.12.0
+
+## 1.11.0
+
+### Minor Changes
+
+- 49c1994: refactor: Deprecate optimism-goerli network
+  chore: Enable zkEVM networks
+  chore: Upgrade npm packages
+
+### Patch Changes
+
+- Updated dependencies [49c1994]
+  - @openzeppelin/defender-sdk@1.11.0
+
+## 1.10.0
+
+### Minor Changes
+
+- 7e14499: fix: Add missing dependencies for cli tool to manage actions.
+  fix: Add confirmation property to tx override.
+  refactor: Remove deprecated config resources
+
+### Patch Changes
+
+- Updated dependencies [7e14499]
+  - @openzeppelin/defender-sdk@1.10.0
+
+## 1.9.0
+
+### Minor Changes
+
+- ed29a58: - chore: Autotask & actions condition names.
+  - feat: Add Web3 & list contract examples.
+  - feat: Add support to Zksync sepolia.
+  - feat: Upgrade dependencies.
+
+### Patch Changes
+
+- Updated dependencies [ed29a58]
+  - @openzeppelin/defender-sdk@1.9.0
+
 ## 1.7.0
 
 ### Minor Changes